\name{var_break_fit}
\alias{var_break_fit}
\title{VAR Model Fitting with LASSO Penalty}
\description{
  Fits a vector autoregression (VAR) model with a LASSO penalty to the given time series data.
}
\usage{
var_break_fit(method, data, lambda, p, initial_phi, max_iteration = 1000, tol = 1e-4, step_size = 1e-3)
}
\arguments{
  \item{method}{A string specifying the method used for fitting. Currently, only "LASSO" is supported.}
  \item{data}{A matrix of time series data, where each column represents a different variable.}
  \item{lambda}{A double value specifying the lambda parameter for the LASSO penalty.}
  \item{p}{An integer specifying the lag order of the VAR model.}
  \item{initial_phi}{An initial value for the phi matrix. If none is provided, a matrix of zeros is used.}
  \item{max_iteration}{An optional integer specifying the maximum number of iterations. Default is 1000.}
  \item{tol}{An optional numeric value indicating the tolerance for convergence. Default is 1e-4.}
  \item{step_size}{An optional numeric value indicating the step size in the optimization algorithm. Default is 1e-3.}
}
\value{
  A list containing:
  \itemize{
    \item{phihat}{A matrix of estimated coefficients for the VAR model.}
    \item{iter}{A matrix indicating the number of iterations taken for each time series variable.}
  }
}
\details{
  The function \code{var_break_fit} fits a VAR model using the LASSO method, which is useful for high-dimensional time series data. The function iteratively updates the coefficients to minimize the penalized least squares criterion.
}
\examples{
# Example usage:
# Generate sample time series data
data <- matrix(rnorm(1000), ncol = 10)
# Fit VAR model with LASSO
result <- var_break_fit("LASSO", data, lambda = 0.5, p = 2)
# View estimated coefficients
print(result$phihat)
}

